Added scripts in __init__ file
