+++ cfme/tests/automate/test_simulation.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
-from textwrap import dedent
-
 import fauxfactory
 import pytest
 
+from . import user_list_hash_data
 from cfme import test_requirements
 from cfme.automate.simulation import simulate
 from cfme.base.ui import AutomateSimulationView
@@ -118,72 +117,7 @@ def test_automate_simulation_result_has_hash_data(custom_instance):
             2.
             3. The UI should display the result objects
     """
-    script = dedent(
-        """
-module Demo
-  module Automate
-    module System
-      module Request
-        module Test
-          class AvailableUsers
-            def initialize(handle = $evm)
-              @handle = handle
-            end
-
-            def main
-              fill_dialog_field(fetch_list_data)
-            end
-
-            private
-
-            def fetch_list_data
-              user1 = {:id => 1, :name => 'Fred', :age => 52, :email => "fred@bedrock.com"}
-              user2 = {:id => 2, :name => 'Barney', :age => 55, :email => "barney@bedrock.com" }
-              user3 = {:id => 3, :name => 'Wilma', :age => 39, :email => "wilma@berdock.com" }
-              user4 = {:id => 4, :name => 'Betty', :age => 38, :email => "betty@bedrock.com" }
-              user_list = { user1 => "#{user1[:name]} - ID #{user1[:id]}",
-                            user2 => "#{user2[:name]} - ID #{user2[:id]}",
-                            user3 => "#{user3[:name]} - ID #{user3[:id]}",
-                            user4 => "#{user4[:name]} - ID #{user4[:id]}"}
-              @handle.log(:info, "User List #{user_list}")
-              return nil => "<none>" if user_list.blank?
-
-              user_list[nil] = "<select>" if user_list.length > 1
-              user_list
-            end
-
-            def fill_dialog_field(list)
-              dialog_field = @handle.object
-
-              # sort_by: value / description / none
-              dialog_field["sort_by"] = "description"
-
-              # sort_order: ascending / descending
-              dialog_field["sort_order"] = "ascending"
-
-              # data_type: string / integer
-              dialog_field["data_type"] = "string"
-
-              # required: true / false
-              dialog_field["required"] = "true"
-
-              dialog_field["values"] = list
-              dialog_field["default_value"] = list.length == 1 ? list.keys.first : nil
-            end
-          end
-        end
-      end
-    end
-  end
-end
-
-if __FILE__ == $PROGRAM_NAME
-  Demo::Automate::System::Request::Test::AvailableUsers.new.main
-end
-        """
-    )
-
-    instance = custom_instance(ruby_code=script)
+    instance = custom_instance(ruby_code=user_list_hash_data)
 
     # Executing automate method
     with LogValidator(
@@ -204,14 +138,12 @@ end
     view = instance.create_view(AutomateSimulationView)
     assert (
         view.result_tree.click_path(
-            *(
-                f"ManageIQ/SYSTEM / PROCESS / {instance.klass.name}",
-                f"ManageIQ/System / {instance.klass.name} / Call_Instance",
-                f"{instance.domain.name}/System / {instance.klass.name} / {instance.name}",
-                "values",
-                "Hash",
-                "Key",
-            )
+            f"ManageIQ/SYSTEM / PROCESS / {instance.klass.name}",
+            f"ManageIQ/System / {instance.klass.name} / Call_Instance",
+            f"{instance.domain.name}/System / {instance.klass.name} / {instance.name}",
+            "values",
+            "Hash",
+            "Key",
         ).text
         == "Key"
     )
